requirements.txt
websockets
python-gsmmodem-new
requests
